Cleans up file names and adds `buildDomain` util
